+++ src/components/letter/LetterUtils.ts
@@ -1,0 +1,7 @@
+/** Returns if the frontend app is in test mode (and therefore should be talking to test stripe checkout flow)
+ *
+ * @return {boolean} is in test mode
+ */
+export const isTestMode = (): boolean => {
+  return window.location.host !== "mail-your-rep.web.app";
+};
